another type missmatch

+++ eufs_rqt/src/eufs_rqt/EUFSRobotSteeringGUI.py
@@ -355,7 +355,7 @@ class EUFSRobotSteeringGUI(Plugin):
 
     def _on_min_linear_changed(self, value):
         self._widget.linear_slider.setMinimum(
-            value * EUFSRobotSteeringGUI.slider_factor)
+            int(value * EUFSRobotSteeringGUI.slider_factor))
 
     def _on_max_angular_changed(self, value):
         self._widget.angular_slider.setMaximum(
